requirements.txt
psutil>=5.9.0
paho-mqtt>=1.6.0